fix(shmem): share pointer_type hash patch across backends
