+++ python/triton_dist/language/extra/ascend/libaclshmem_device.py
@@ -3,6 +3,7 @@
 from triton.language import core
 import triton.language as tl
 from triton_dist.language.core import extern_call
+from triton_dist.language.extra.utils import patch_hash_method_for_pointer_type
 
 pi_u64_t = tl.core.pointer_type(tl.core.dtype("uint64"))
 pi_i32_t = tl.core.pointer_type(tl.core.dtype("int32"))
@@ -38,20 +39,6 @@ DTYPE_TO_KERNEL_SUFFIX = {
     "bf16": "bfloat16",
 }
 
-
-def _pointer_type_hash(self):
-    return hash((self.name, self.element_ty, "tt_ptr"))
-
-
-def patch_hash_method_for_pointer_type():
-    elem_dtype_list = (tl.core.dtype.SINT_TYPES + tl.core.dtype.UINT_TYPES + tl.core.dtype.FP_TYPES +
-                       tl.core.dtype.OTHER_TYPES)
-    for elem_dtype in elem_dtype_list:
-        ptr_ty = type(tl.core.pointer_type(tl.core.dtype(elem_dtype)))
-        ptr_ty.__hash__ = _pointer_type_hash
-
-
-# Keep pointer_type hashable for Triton extern signature dict keys.
 patch_hash_method_for_pointer_type()
